requirements.txt
e2b-code-interpreter>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0